+++ 4Prak/code-with-quarkus/src/main/java/org/mannschaftssport/boundary/ACL/SpielerpassDTO.java
@@ -3,14 +3,14 @@ package org.mannschaftssport.boundary.ACL;
 import javax.ws.rs.core.Link;
 
 public class SpielerpassDTO {
-    public long id;
+    public int id;
     public String name;
 
     public SpielerpassDTO(){
 
     }
 
-    public SpielerpassDTO(long id, String name){
+    public SpielerpassDTO(int id, String name){
         this.id = id;
         this.name = name;
     }
